add inventory feature for open orders
Inventory always shows the state of shipped orders
In other words, shipping orders reduces their inventory

But sometimes orders stay open for a while and it is helpful to know
the
inventory if those orders were shipped. So the previous commit made a
button for that and this one does the functionality.
(no tests though, shame on dancinglightning)
open orders must be less than a month old.